requirements.txt
appdirs==1.4.3
atlassian-python-api==1.17.5
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
html2text==2020.1.16
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
pipdeptree==1.0.0
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
webencodings==0.5.1
